not sure how to implement a Data Flip-Flop in this simulator
It would be nice to learn how to actually implement a DFF rather than rely
upon the built-in one. However, I wonder if it is possible using this
simulator.

Also, I need to take a closer look at the docs for the CLOCKED keyword
because so far no syntax works.